requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
openai==1.57.2
pyyaml==6.0.2
httpx==0.27.2
